% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_understoreys.R
\name{assembly_understoreys}
\alias{assembly_understoreys}
\title{Title}
\usage{
assembly_understoreys(trollsim_pre, trollsim_post, shape = "TotalDamages",
  coords = c("Xutm", "Yutm"), plotnum, year, isUTM = T,
  path_graph = "C:/Users/nino.page/Desktop/TROLL project/Paracou/damages/graphs/leaflets",
  pathshp = "C:/Users/nino.page/Desktop/TROLL project/Paracou/damages/results",
  pathlim = "C:/Users/nino.page/Desktop/TROLL project/Paracou/limits")
}
\arguments{
\item{trollsim_pre}{data.frame, inventory from trollsim pre-logging for mature understoreys}

\item{trollsim_post}{data.frame, inventory from trollsim post-logging for immature understoreys}

\item{shape}{character. Name of the shapefile used for damage without any extension.}

\item{pathlim}{}
}
\description{
Title
}
