% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damage_logging.R
\name{damage_logging}
\alias{damage_logging}
\title{Title}
\usage{
damage_logging(trollsim, shape = "TotalDamages", coords = c("Xutm", "Yutm"),
  plotnum, year, isUTM = T,
  path_graph = "C:/Users/nino.page/Desktop/TROLL project/Paracou/damages/graphs/leaflets",
  pathshp = "C:/Users/nino.page/Desktop/TROLL project/Paracou/damages/results",
  pathlim = "C:/Users/nino.page/Desktop/TROLL project/Paracou/limits")
}
\arguments{
\item{trollsim}{data.frame - Forest simulated with TROLL or trollsim object}

\item{isUTM}{}
}
\description{
Title
}
