% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractplot.R
\name{extractplot}
\alias{extractplot}
\title{Title}
\usage{
extractplot(data, plotcol = "n_parcelle", plot = 11,
  sizecol = "circ_corr", centimeters = T, alivecol = "code_vivant",
  circonference = T, Genre = "Genre", Espece = "Espece",
  species_name = FALSE, Paracou_correct = TRUE)
}
\arguments{
\item{data}{data.frame Paracou dataset, for example}

\item{plotcol}{Column indicating the plot IDs, that are ideally integers}

\item{plot}{Integer. The number corresponding to the plot you wanna extract}

\item{sizecol}{Character. The column indicating the tree size measurement : 
DBH or circonference. The latter will need to be converted, which is done if "circonference = T"}

\item{centimeters}{Logical. If your DBH are in centimeters, function converts it into milimeters.}

\item{alivecol}{Character. Column indicating if measured trees are alive, 
dead, fallen, etc. Only live trees are retained.}

\item{circonference}{Logical. If true, conversion to dbh will be done. 
If false, you mean that size measurements in your data are diameters at breast height.}

\item{Genre}{Character. The column indicating the individuals' genera.}

\item{Espece}{Character. The column indicating the species names.}

\item{species_name}{Logical. False, if you have genus and species
name separatedly. True if you already have a column aggregating genus and species names 
separated by an under_score.}

\item{Paracou_correct}{Logical. Specific of Paracou, to prevent crashes from database-inherent singularities}
}
\value{
Output is a dataframe corresponding to all the inventories for the plot you selected,
 with species column added and circonference converted to dbh if ever needed.
}
\description{
Title
}
