% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_div_species.R
\name{fonctional_div_species}
\alias{fonctional_div_species}
\title{Troll Model Validation : functional diversity with Leinster-Cobbold Entropy
using resampling and "rolled under the armpits" Goodness of Fit statistics.}
\usage{
fonctional_div_species(data, trollsim, speciescol, speciestraits, LMA = "LMA",
  N = "N", P = "P", wsg = "wsg", ah = "ah", hmax = "hmax",
  dmax = "dmax", n_resample = 10000, path, nameplot)
}
\arguments{
\item{data}{Data.frame ; An inventory which contains at least species name.}

\item{trollsim}{Data.frame (yet not adapted to trollsim S4 objects)with species name
Diameters should correspon in terms of lower bound (generally 10cm) between
field data and trollsim : pre-filtering is recommended to make consistent comparisons.}

\item{speciescol}{Character indicating the name of the species identity column}

\item{speciestraits}{Data.frame the contains species name and the 7 traits (species-specific mean values)}

\item{LMA}{Character indicating the name of the LMA species means column}

\item{N}{Character indicating the name of the N species means column}

\item{P}{Character indicating the name of the P species means column}

\item{wsg}{Character indicating the name of the wsg species means column}

\item{ah}{Character indicating the name of the ah species means column}

\item{hmax}{Character indicating the name of the hmax species means column}

\item{dmax}{Character indicating the name of the dmax species means column}

\item{n_resample}{Integer ; the number of times you wish to resample to infer sampling effects.}

\item{path}{Character ; The path to the directory you want to save the graphical outputs into.}

\item{nameplot}{Character ; the name of your plot (used to generate outputs' filenames)}
}
\value{
A matrix of computed distances of the resampled Leinster-Cobbold Entropy profiles (LCEP)
and the original dataset's LCEP using two distance statistics : (is the first one alright ??)
    dist = sqrt(sum((LC_data - LC_sample)^2)/nrow(LC_data)) and chi =(sum((LC_sample-LC_data)^2)/LC_data)
}
\description{
Troll Model Validation : functional diversity with Leinster-Cobbold Entropy
using resampling and "rolled under the armpits" Goodness of Fit statistics.
}
