% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_general.R
\name{init_general}
\alias{init_general}
\title{General and Species Parameters Input-File Maker for TROLL Forest Simulaor}
\usage{
init_general(macros = "intra", path, input, overwrite = TRUE, nbcol = 400,
  nbrows = 400, nbiter = 12, iter = 12, NV = 1, NH = 1, nbout = 4,
  numesp = 163, p = 0.05, klight = 0.9, phi = 0.06, g1 = 3.77,
  vC = 0.1, DBH0 = 0.01, H0 = 1, ra0 = 0.5, ra1 = 10, de0 = 0.3,
  de1 = 0.12, dens = 0.8, fbranchstem = 0.35, fcanopy = 0.3,
  seedrain = 5000, nbseeds = 10, sigma_height = 0.05, sigma_CR = 0.05,
  sigma_CD = 0.05, sigma_P = 0.05, sigma_N = 0.05, sigma_LMA = 0.05,
  sigma_wsg = 0.05, sigma_dmax = 0.05, corr_CR_height = 0.75,
  corr_N_P = 0.65, corr_N_LMA = -0.43, corr_P_LMA = -0.39,
  leafdem_resolution = 30, p_tfsecondary = 0, hurt_decay = 0,
  mindeathrate = 0.035, m1 = 0.035, CO2 = 360, species)
}
\arguments{
\item{macros}{Character. MUST BE "species-specific", "intraspecific-only" or "intraspecific-covar".
Describes the macros that you use for this simulation, so that the function adapts the input file to avoid crashing TROLL.}

\item{path}{Character - The path leading to the directory you want to save the file in.}

\item{input}{Character - The name of the file}

\item{overwrite}{Logical - TRUE if you want to overwrite file existing with same name.}

\item{nbcol}{- Integer, dimensions of your grid (number of cells)}

\item{nbrows}{-Integer, dimensions of your grid (number of cells)}

\item{NV}{- Integer, number of vertical cells per meter (m-1)}

\item{NH}{- Integer, number of horizontal cells (m-2)}

\item{numesp}{- Integer, number of species}

\item{species}{}
}
\description{
General and Species Parameters Input-File Maker for TROLL Forest Simulaor
}
