% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_out.R
\name{inv_out}
\alias{inv_out}
\title{Title}
\usage{
inv_out(x, forest = getOption("RconTroll.forest"),
  path = getOption("RconTroll.path"), overwrite = TRUE)
}
\arguments{
\item{overwrite}{}
}
\description{
Title
}
