% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_out.R
\name{load_out}
\alias{load_out}
\title{Title}
\usage{
load_out(name = getOption("RconTroll.name"),
  path = file.path(getOption("RconTroll.path"), getOption("RconTroll.name")))
}
\arguments{
\item{path}{}
}
\description{
Title
}
