% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareforest.R
\name{prepareforest}
\alias{prepareforest}
\title{Title}
\usage{
prepareforest(data, year = 1992, censuscol = "campagne", label = "sp_lab",
  speciescol = "species", speciestraits, replacement = "local",
  i_arbre = "i_arbre", X = "X", Y = "Y", dbh = "dbh", milimeters = T,
  plotcol = "n_parcelle")
}
\arguments{
\item{data}{data.frame ; Basically the censuses for the plot you want to work on.}

\item{year}{Integer. The year corresponding to your simulation startpoint.}

\item{censuscol}{Character. Name of the column indicating the censuses years.}

\item{label}{Character.Name of the column indicating species labels, if you have defined them.}

\item{speciescol}{Character. The name of the column indicating species identity. Should be of the form Genus_species matching case and separator.}

\item{speciestraits}{Data.frame. The table containing your species-specific traits that will be used to parametrise TROLL.}

\item{replacement}{Character. Procedure you choose to replace individuals belonging to missing species.}

\item{X}{Character. The name of the column containing X coordinates. At the plot scale, don't put an UTM or other shit !}

\item{Y}{Character. The name of the column containing Y coordinates. At the plot scale, don't put an UTM or other shit !}

\item{dbh}{Character. Name of the column containing the diameters at breast height. Circumferences are handler in the "extractplot" function.}

\item{milimeters}{Logical. Indicates whether the diameters are in milimeters or not.
Defaults to FALSE as this unit is unusual in most censuses.}
}
\description{
Title
}
