% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplot.R
\name{speciesplot}
\alias{speciesplot}
\title{Title}
\usage{
speciesplot(dataplot, speciescol = "species",
  sp = read.table("C:/Users/nino.page/Desktop/TROLL project/data/new_species.txt",
  header = T))
}
\arguments{
\item{dataplot}{Data.frame... Data obtained from extractplot()}

\item{speciescol}{Character. name of the species column}

\item{sp}{species-traits dataset..}
}
\value{
A dataframe with the plot's species that match with TROLL sp list
and the corresponding "regional" frequencies to calibrate the seedrain on.
}
\description{
Title
}
