% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesstation.R
\name{speciesstation}
\alias{speciesstation}
\title{Title}
\usage{
speciesstation(datastation, speciescol = "species",
  sp = read.table("species.txt", header = T))
}
\arguments{
\item{datastation}{whole dataset of your station...}

\item{speciescol}{Character. name of the species column}

\item{sp}{species-traits dataset..}

\item{dataplot}{Data.frame... Data obtained from extractplot()}
}
\description{
Title
}
