% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understorey.R
\name{understorey}
\alias{understorey}
\title{Title}
\usage{
understorey(realdata, trollsim, plotnum, X = "X", Y = "Y", dbh = "dbh",
  sp_lab = "sp_lab", path)
}
\arguments{
\item{path}{}
}
\description{
Title
}
